reader: fix handling of sizes in general_work
Currently, the general_work implementation has the following issues
that affect specially files with repeat=False:
* After all the GUPPI blocks have been read, general_work() will
  produce 0 items, even if there are still samples in the buffer.
  Consequently these samples get lost.
* general_work() never returns WORK_DONE, so a flowgraph that reads
  from a file with repeat=False will never stop.

This fixes these problems by restructuring general_work() somewhat.
If the buffer is not full, we try to read a GUPPI block. If the file
hasn't finished, we put the block in the buffer. Otherwise, in repeat
mode we restart the file. Then we produce the maximum number of samples
according to the buffer occupancy and length of output_items. If there
is nothing to produce and repeat mode is disabled, we finish the work.